+++ lib/src/chart/base/axis_chart/axis_chart_data.dart
@@ -1116,13 +1116,14 @@ class VerticalLine extends FlLine with EquatableMixin {
 /// Draws a title on the [HorizontalLine]
 class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
   /// Draws a title on the [HorizontalLine], align it with [alignment] over the line,
-  /// applies [padding] for spaces, and applies [style for changing color,
+  /// applies [padding] for label padding, applies [margin] for label space andapplies [style] for changing color,
   /// size, ... of the text.
   /// Drawing text will retrieve through [labelResolver],
   /// you can override it with your custom data.
   /// /// [show] determines showing label or not.
   HorizontalLineLabel({
     EdgeInsets? padding,
+    EdgeInsets? margin,
     super.style,
     Alignment? alignment,
     super.show = false,
@@ -1130,7 +1131,8 @@ class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
   })  : labelResolver =
             labelResolver ?? HorizontalLineLabel.defaultLineLabelResolver,
         super(
-          padding: padding ?? const EdgeInsets.all(6),
+          padding: padding ?? EdgeInsets.zero,
+          margin: margin ?? EdgeInsets.zero,
           alignment: alignment ?? Alignment.topLeft,
         );
 
@@ -1163,6 +1165,7 @@ class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
         labelResolver,
         show,
         padding,
+        margin,
         style,
         alignment,
       ];
@@ -1171,13 +1174,14 @@ class HorizontalLineLabel extends FlLineLabel with EquatableMixin {
 /// Draws a title on the [VerticalLine]
 class VerticalLineLabel extends FlLineLabel with EquatableMixin {
   /// Draws a title on the [VerticalLine], align it with [alignment] over the line,
-  /// applies [padding] for spaces, and applies [style for changing color,
+  /// applies [padding] for label padding, applies [margin] for label space andapplies [style] for changing color,
   /// size, ... of the text.
   /// Drawing text will retrieve through [labelResolver],
   /// you can override it with your custom data.
   /// [show] determines showing label or not.
   VerticalLineLabel({
     EdgeInsets? padding,
+    EdgeInsets? margin,
     TextStyle? style,
     Alignment? alignment,
     bool? show,
@@ -1186,7 +1190,8 @@ class VerticalLineLabel extends FlLineLabel with EquatableMixin {
             labelResolver ?? VerticalLineLabel.defaultLineLabelResolver,
         super(
           show: show ?? false,
-          padding: padding ?? const EdgeInsets.all(6),
+          padding: padding ?? EdgeInsets.zero,
+          margin: margin ?? EdgeInsets.zero,
           style: style ??
               const TextStyle(
                 color: Colors.black,
@@ -1212,6 +1217,8 @@ class VerticalLineLabel extends FlLineLabel with EquatableMixin {
     return VerticalLineLabel(
       padding:
           EdgeInsets.lerp(a.padding as EdgeInsets, b.padding as EdgeInsets, t),
+      margin:
+          EdgeInsets.lerp(a.margin as EdgeInsets, b.margin as EdgeInsets, t),
       style: TextStyle.lerp(a.style, b.style, t),
       alignment: Alignment.lerp(a.alignment, b.alignment, t),
       labelResolver: b.labelResolver,
@@ -1225,6 +1232,7 @@ class VerticalLineLabel extends FlLineLabel with EquatableMixin {
         labelResolver,
         show,
         padding,
+        margin,
         style,
         alignment,
       ];