+++ lib/src/extensions/offset_extension.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/widgets.dart';
+
+extension OffsetExtension on Offset {
+  Offset applyEdgeInsets(EdgeInsets inset) {
+    return Offset(
+      dx - inset.right + inset.left,
+      dy - inset.bottom + inset.top,
+    );
+  }
+}